-chg: This version incorporates the changes made by scite-ru (internationalization unicode fix) and definitely builds with mingw! I had to add a few defines which weren't in my mingw headers, and a few defines which weren't in my old VC6 headers ;)
The other fix is that floating toolbars no longer grab the focus, and
they are correctly reshown in the mingw version (apparently DllMain is
not called, which is why I have an explicit initialization call in
luaopen_gui)

There are a lot of requested features, which I must go through. One
thing that Frank mentioned is that it would be nice to hide/show the
side panel.

steve d.